Move withCursors to pipeline helpers
